Use repo_bootstrap to break dependency cycle
There is a dependency cycle between Sahara and its plugins. If building
on CentOS 7, the api and engine subpackages depend on the plugins,
which depend on python2-sahara.

We need to use repo_bootstrap to break this dependency cycle, so
we can first get a working sahara package, then build its plugins.

Change-Id: I0dfd80a02b2d348ad919c1242b00b24ad15ffefe